fix:can't sync binlog when exec redis cmd setnx (#940)
bugfix: can't sync binlog when exec redis command setnx and the key was exist.
when exec cmd setnx, if the key was exist, the value was not update by this cmd.
the binlog of this cmd will be saved into binlog file, but binlog length was zero. so can't do sync log.
so i fixed this issue, by writing binlog of setnx not judge from command return value.